+++ examples/examples_chain/src/bin/async_chain_ops.rs
@@ -1,4 +1,5 @@
 use optionstratlib::prelude::*;
+use osl_example_support::setup_logger;
 use std::error::Error;
 
 #[tokio::main]