+++ examples/examples_simulation/src/bin/position_simulator.rs
@@ -1,4 +1,5 @@
 use optionstratlib::prelude::*;
+use osl_example_support::setup_logger;
 use positive::pos_or_panic;
 
 #[derive(Clone)]